#95 [FEAT] : isSelect 및 visibility DataBinding 적용
