+++ app/src/main/java/com/example/charo_android/presentation/ui/BindingConversions.kt
@@ -1,0 +1,73 @@
+package com.example.charo_android.presentation.ui
+
+import android.util.DisplayMetrics
+import android.util.Log
+import android.view.View
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.charo_android.R
+
+object BindingConversions {
+    @BindingAdapter("profileImageUrl")
+    @JvmStatic
+    fun loadMyPageProfileImage(imageView: ImageView, url: String?) {
+        if(url != null) {
+            Glide.with(imageView.context)
+                .load(url)
+                .override(56 * DisplayMetrics().densityDpi, 56 * DisplayMetrics().densityDpi)
+                .circleCrop()
+                .into(imageView)
+        }
+    }
+
+    @BindingAdapter("myPageRecyclerViewImageUrl")
+    @JvmStatic
+    fun loadMyPgaeRecyclerViewImage(imageView: ImageView, url: String?) {
+        if(url != null) {
+            Glide.with(imageView.context)
+                .load(url)
+                .transform(RoundedCorners(9))
+                .into(imageView)
+        }
+    }
+
+    @BindingAdapter("detailViewProfileImageUrl")
+    @JvmStatic
+    fun loadDetailProfileImage(imageView: ImageView, url: String?) {
+        if(url != null) {
+            Glide.with(imageView.context)
+                .load(url)
+                .override(29 * DisplayMetrics().densityDpi, 29*DisplayMetrics().densityDpi)
+                .circleCrop()
+                .into(imageView)
+        }
+    }
+
+    @BindingAdapter("is_selected")
+    @JvmStatic
+    fun setSelected(view: View, selected: Boolean) {
+        view.isSelected = selected
+    }
+
+    @BindingAdapter("is_INVISIBLE")
+    @JvmStatic
+    fun setInvisible(view: View, flag: Boolean) {
+        if(flag) {
+            view.visibility = View.INVISIBLE
+        } else {
+            view.visibility = View.VISIBLE
+        }
+    }
+
+    @BindingAdapter("is_GONE")
+    @JvmStatic
+    fun setGone(view: View, flag: Boolean) {
+        if(flag) {
+            view.visibility = View.GONE
+        } else {
+            view.visibility = View.VISIBLE
+        }
+    }
+}
